Pyhon-Services/requirements.txt
fastapi
uvicorn
python-barcode
fpdf2
pillow
reportlab